Switch dates from i128 to u64
I was making them unnecessarily large by accidentally storing them as microseconds instead of milliseconds. Might be able to get away with seconds in the future

+++ src/types.zig
@@ -2,7 +2,7 @@ pub const ImportedScrobble = struct {
     track: []const u8,
     artist: []const u8,
     album: []const u8 = "",
-    date: i128,
+    date: u64,
 };
 
 pub const Scrobble = struct {
@@ -10,7 +10,7 @@ pub const Scrobble = struct {
     artists_track: []const []const u8,
     album: []const u8 = "",
     artists_album: []const []const u8,
-    date: i128,
+    date: u64,
 };
 
 // From lastfmstats.com
